Fix code definitions in __init__.py (ndfourier)
